Improvements to start form handling
