+++ src/management-system-v2/app/(dashboard)/[environmentId]/(automation)/executions/[processId]/instance-variables.tsx
@@ -1,31 +1,22 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useState } from 'react';
 import { RelevantInstanceInfo } from './instance-info-panel';
 
 import { EditOutlined } from '@ant-design/icons';
 
-import type { Variable as ProcessVariable } from '@proceed/bpmn-helper/src/getters';
-import { getProcessIds, getVariablesFromElementById } from '@proceed/bpmn-helper';
 import { App, Button, Form, Input, InputNumber, Modal, Switch, Table } from 'antd';
 import { updateVariables } from '@/lib/engines/server-actions';
 import { useEnvironment } from '@/components/auth-can';
 import TextArea from 'antd/es/input/TextArea';
 import { wrapServerCall } from '@/lib/wrap-server-call';
 import { typeLabelMap } from '../../../processes/[processId]/use-process-variables';
+import useInstanceVariables, { Variable } from './use-instance-variables';
 
 type InstanceVariableProps = {
   info: RelevantInstanceInfo;
   refetch: () => void;
 };
 
-type Variable = {
-  name: string;
-  type: 'string' | 'number' | 'boolean' | 'object' | 'array' | 'unknown';
-  allowed?: string;
-  value: any;
-};
-
 const InstanceVariables: React.FC<InstanceVariableProps> = ({ info, refetch }) => {
-  const [variableDefinitions, setVariableDefinitions] = useState<ProcessVariable[]>([]);
   const [updatedValue, setUpdatedValue] = useState<any>(undefined);
   const [submitting, setSubmitting] = useState(false);
 
@@ -35,67 +26,7 @@ const InstanceVariables: React.FC<InstanceVariableProps> = ({ info, refetch }) =
 
   const [form] = Form.useForm();
 
-  useEffect(() => {
-    const initVariables = async () => {
-      const [processId] = await getProcessIds(info.version.bpmn);
-      const variables = await getVariablesFromElementById(info.version.bpmn, processId);
-      setVariableDefinitions(
-        variables.map((v) => ({
-          ...v,
-        })),
-      );
-    };
-    initVariables();
-  }, [info.version]);
-
-  const variables = useMemo(() => {
-    const { instance } = info;
-
-    const variables: Record<string, Variable> = {};
-
-    if (instance) {
-      const instanceVariables = instance.variables as Record<string, { value: any }>;
-
-      Object.entries(instanceVariables).forEach(([name, { value }]) => {
-        let type: Variable['type'] = 'unknown';
-        const valueType = typeof value;
-        switch (valueType) {
-          case 'number':
-          case 'boolean':
-          case 'string':
-            type = valueType;
-            break;
-          case 'object': {
-            if (Array.isArray(value)) {
-              type = 'array';
-              break;
-            } else if (value) {
-              type = 'object';
-              break;
-            }
-          }
-        }
-
-        variables[name] = { name, type, value };
-      });
-    }
-
-    variableDefinitions.forEach((def) => {
-      if (!variables[def.name]) {
-        variables[def.name] = {
-          name: def.name,
-          type: def.dataType as Variable['type'],
-          value: undefined,
-        };
-      }
-      if (variables[def.name].type === 'unknown') {
-        variables[def.name].type = def.dataType as Variable['type'];
-      }
-      variables[def.name].allowed = def.enum;
-    });
-
-    return Object.values(variables);
-  }, [variableDefinitions, info.instance]);
+  const { variables } = useInstanceVariables(info);
 
   const [variableToEdit, setVariableToEdit] = useState<Variable | undefined>(undefined);
 
@@ -263,25 +194,23 @@ const InstanceVariables: React.FC<InstanceVariableProps> = ({ info, refetch }) =
                         break;
                       // check if the entered text can be transformed to the respective object type
                       case 'object':
-                      case 'array':
-                        {
-                          try {
-                            const parsed = JSON.parse(value);
-                            if (
-                              (variableToEdit.type === 'array' && Array.isArray(parsed)) ||
-                              (variableToEdit.type === 'object' && !Array.isArray(parsed))
-                            ) {
-                              return Promise.resolve();
-                            }
-                          } catch (err) {}
+                      case 'array': {
+                        try {
+                          const parsed = JSON.parse(value);
+                          if (
+                            (variableToEdit.type === 'array' && Array.isArray(parsed)) ||
+                            (variableToEdit.type === 'object' && !Array.isArray(parsed))
+                          ) {
+                            return Promise.resolve();
+                          }
+                        } catch (err) {}
 
-                          return Promise.reject(
-                            new Error(
-                              `Input is not a valid JSON ${variableToEdit.type === 'object' ? 'Object' : 'List'}.`,
-                            ),
-                          );
-                        }
-                        break;
+                        return Promise.reject(
+                          new Error(
+                            `Input is not a valid JSON ${variableToEdit.type === 'object' ? 'Object' : 'List'}.`,
+                          ),
+                        );
+                      }
                     }
                   }
