+++ src/management-system-v2/lib/engines/tasklist.ts
@@ -2,7 +2,6 @@
 
 import { Engine } from './machines';
 import { engineRequest } from './endpoints/index';
-import { inlineScript, inlineUserTaskData } from '@proceed/user-task-helper';
 
 export type TaskListEntry = {
   id: string;
@@ -82,14 +81,6 @@ export async function getStartFormFromMachine(
       version: versionId,
     },
   });
-
-  // initialize the placeholders in the form with empty strings
-  // TODO: use the information from the variable data in the bpmn to initialize the actual initial
-  // values set by the process designer
-  if (html) {
-    html = inlineScript(html, '', '');
-    html = inlineUserTaskData(html, {}, []);
-  }
 
   return html as string;
 }
